+++ pc/delay_pc.c
@@ -47,9 +47,9 @@
  * \param ms Anzahl der Millisekunden
  */
 void delay(uint16_t ms) {
-	int wait = ms;
-	int steps = wait/999; // max sleep-time = 10^6-1 us
-	int i;
+	int16_t wait = ms;
+	int16_t steps = wait/999; // max sleep-time = 10^6-1 us
+	int16_t i;
 	/* in 999 ms Schritten warten */
 	for (i=0; i<steps; i++) {
 		SLEEP(999);
